First pass on update to 0.7.2
