+++ dl4j-examples/src/main/java/org/deeplearning4j/examples/misc/customlayers/layer/CustomLayer.java
@@ -6,6 +6,8 @@ import org.deeplearning4j.nn.conf.NeuralNetConfiguration;
 import org.deeplearning4j.nn.conf.layers.FeedForwardLayer;
 import org.deeplearning4j.nn.params.DefaultParamInitializer;
 import org.deeplearning4j.optimize.api.IterationListener;
+import org.nd4j.linalg.activations.Activation;
+import org.nd4j.linalg.activations.IActivation;
 import org.nd4j.linalg.api.ndarray.INDArray;
 
 import java.util.Collection;
@@ -18,7 +20,7 @@ import java.util.Map;
  */
 public class CustomLayer extends FeedForwardLayer {
 
-    private String secondActivationFunction;
+    private IActivation secondActivationFunction;
 
     public CustomLayer() {
         //We need a no-arg constructor so we can deserialize the configuration from JSON or YAML format
@@ -31,12 +33,12 @@ public class CustomLayer extends FeedForwardLayer {
         this.secondActivationFunction = builder.secondActivationFunction;
     }
 
-    public String getSecondActivationFunction() {
+    public IActivation getSecondActivationFunction() {
         //We also need setter/getter methods for our layer configuration fields (if any) for JSON serialization
         return secondActivationFunction;
     }
 
-    public void setSecondActivationFunction(String secondActivationFunction) {
+    public void setSecondActivationFunction(IActivation secondActivationFunction) {
         //We also need setter/getter methods for our layer configuration fields (if any) for JSON serialization
         this.secondActivationFunction = secondActivationFunction;
     }
@@ -83,7 +85,7 @@ public class CustomLayer extends FeedForwardLayer {
     //Note that we are inheriting all of the FeedForwardLayer.Builder options: things like n
     public static class Builder extends FeedForwardLayer.Builder<Builder> {
 
-        private String secondActivationFunction;
+        private IActivation secondActivationFunction;
 
         //This is an example of a custom property in the configuration
 
@@ -93,6 +95,15 @@ public class CustomLayer extends FeedForwardLayer {
          * @param secondActivationFunction Second activation function for the layer
          */
         public Builder secondActivationFunction(String secondActivationFunction) {
+            return secondActivationFunction(Activation.fromString(secondActivationFunction).getActivationFunction());
+        }
+
+        /**
+         * A custom property used in this custom layer example. See the CustomLayerExampleReadme.md for details
+         *
+         * @param secondActivationFunction Second activation function for the layer
+         */
+        public Builder secondActivationFunction(IActivation secondActivationFunction){
             this.secondActivationFunction = secondActivationFunction;
             return this;
         }
