+++ pom.xml
@@ -14,10 +14,10 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <shadedClassifier>bin</shadedClassifier>
         <java.version>1.8</java.version>
-        <nd4j.version>0.7.2-SNAPSHOT</nd4j.version>
-        <dl4j.version>0.7.2-SNAPSHOT</dl4j.version>
-        <datavec.version>0.7.2-SNAPSHOT</datavec.version>
-        <arbiter.version>0.7.2-SNAPSHOT</arbiter.version>
+        <nd4j.version>0.7.2</nd4j.version>
+        <dl4j.version>0.7.2</dl4j.version>
+        <datavec.version>0.7.2</datavec.version>
+        <arbiter.version>0.7.2</arbiter.version>
         <guava.version>19.0</guava.version>
         <logback.version>1.1.7</logback.version>
         <jfreechart.version>1.0.13</jfreechart.version>
